Move `codeMirror` page object to `tests/pages/components`
