+++ tests/pages/course-page.js
@@ -15,7 +15,7 @@ import SecondStageInstructionsCard from 'codecrafters-frontend/tests/pages/compo
 import Sidebar from 'codecrafters-frontend/tests/pages/components/course-page/sidebar';
 import TestResultsBar from 'codecrafters-frontend/tests/pages/components/course-page/test-results-bar';
 import YourTaskCard from 'codecrafters-frontend/tests/pages/components/course-page/course-stage-step/your-task-card';
-import codeMirror from 'codecrafters-frontend/tests/support/code-mirror-page-object';
+import codeMirror from 'codecrafters-frontend/tests/pages/components/code-mirror';
 import { collection, clickable, clickOnText, create, fillable, isVisible, text, triggerable, visitable, hasClass } from 'ember-cli-page-object';
 
 export default create({
